Move update_protocol_document method from CloseMeetingCommand to meeting itself.
